Added props.items listener and item set/restore/replace and save
Added to action creators, types and reducers
Up[dated HOC to listen for props.items changes. If the Dev provides a new list - it will replace and the current list and essentially start over.
This assumes that the dev is controlling items state from outside SelecThis. Otherwise we handle it internally.