Corrected Clear Logic and Changed to Reset
Reset is more accurate to what the clear button is doing, it returns it to the original state of a multi-select

Also groomed the item click logic to us map ensuring purity of object, no side effects. New state for reducer is a new clean object. This also manages subitems when they are present.

Changed isNotItemClicked to isItemClicked - more logical/intuitive . Just use !isItemClicked for NOT, that it what its for.